Update group_project.js (again)
Okay this time I added the bcc, which is only on the student side because the student shouldn't see the people bcc'd by the admin and I'm sure the admin knows what bcc is, so why give them it? The admin side doesn't get the bcc data either; it's stored with the regular email array and can only be seen by going into student_view_sent. I also added a function to reset the help popup data